+++ modulo3/projeto-labeddit/labeddit/src/components/PostCard/styled.js
@@ -1,7 +1,10 @@
 import styled from "styled-components"
 
 export const PostCardContainer = styled.div`
-
+    display: flex;
+    flex-direction: column;
+    width: 100%;
+    margin-bottom: 3%;
 `
 
 export const PostCardContent = styled.div`
