+++ modulo3/projeto-labeddit/labeddit/src/pages/FeedPage/styled.js
@@ -1,5 +1,5 @@
 import styled from 'styled-components'
-import { Fab } from '@material-ui/core'
+
 
 export const FeedPageComtainer = styled.div`
   display: flex;
@@ -7,12 +7,6 @@ export const FeedPageComtainer = styled.div`
   justify-content: center;
   align-items: center;
   height: 100%;
-  margin: 30px;
+  margin: 20px;
 `
 
-export const AddPostButton = styled(Fab)`
-  position: fixed !important;
-  right: 20px;
-  bottom: 20px;
-  z-index: 3;
-`
