+++ modulo3/projeto-labeddit/labeddit/src/pages/PostPage/styled.js
@@ -1,4 +1,5 @@
 import styled from "styled-components"
+import { Fab } from '@material-ui/core'
 
 export const ScreenContainer = styled.div`
   display: flex;
@@ -13,4 +14,11 @@ export const PostContainer = styled.div`
   width: 60vw;
   max-width: 600px;
   min-width: 300px;
+`
+
+export const AddPostButton = styled(Fab)`
+  position: fixed !important;
+  right: 20px;
+  bottom: 20px;
+  z-index: 3;
 `
